Extract logic for 1 loop from Game to GameLooper
